+++ trunk/GoJava7/a.sukhenko/Kickstarter/src/main/java/ua/com/goit/gojava7/kickstarter/dao/storage/ProjectStorage.java
@@ -11,7 +11,7 @@ public interface ProjectStorage extends Storage<Project>{
 
 	public void updatePledged(Project project, int amount);
 
-	public int getPledged(String projectName);
+	public double getPledged(String projectName);
 
 	public List<Project> getByCategory(String categoryName);
